% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatitos.R
\name{gatitos}
\alias{gatitos}
\title{Una función espléndida que transforma todo en gatito}
\usage{
gatitos(x)
}
\arguments{
\item{x}{Cualquier objeto}
}
\value{
Un gatito
}
\description{
Una función espléndida que transforma todo en gatito
}
\examples{
frutas <- c("manzana", "naranja", "pera")
gatitos(frutas)
}
